Implement web-based SSH for serial console on OpenVZ.
When a user requests console access (via the VNC button), we grab the IP/username/password and associate it with a random token. The token is provided to the user, who is then redirected to the wssh server (running on a different port) via iframe. The user sends the token as cookie to the websocket connection, and a tunnel is created to the SSH server.

Screen resizing doesn't work yet; hardcoded 80 by 40 characters.

This commit also fixes issue with console access not being enabled by Lobster in the API call.