websockify: support binary protocol.
In some web server setups, the browser apparently will only request
the binary websocket subprotocol, and not offer base64 support. This
commit adds support for binary subprotocol in the Lobster websockify
module so that the noVNC session can be supported in such setups.

Currently we do not prefer base64 or binary, we simply use whichever
one that appears first in the set of subprotocols offered by the
client. If neither base64 nor binary is offered, we will arbitrarily
pick a subprotocol (this behavior is unchanged from before).